LPD-54244 Create base class to implement getFDSTableSchema method
